WIP - moving debug tools between files.
