+++ src/model/rootelement.js
@@ -98,4 +98,12 @@ export default class RootElement extends Element {
 	toJSON() {
 		return this.rootName;
 	}
+
+	// @if CK_DEBUG_ENGINE // toString() {
+	// @if CK_DEBUG_ENGINE // 	return this.rootName;
+	// @if CK_DEBUG_ENGINE // }
+
+	// @if CK_DEBUG_ENGINE // log() {
+	// @if CK_DEBUG_ENGINE // 	console.log( 'ModelRootElement: ' + this );
+	// @if CK_DEBUG_ENGINE // }
 }
