+++ src/model/text.js
@@ -9,6 +9,8 @@
 
 import Node from './node';
 
+// @if CK_DEBUG_ENGINE // const { convertMapToStringifiedObject } = require( '../dev-tools/utils' );
+
 /**
  * Model text node. Type of {@link module:engine/model/node~Node node} that contains {@link module:engine/model/text~Text#data text data}.
  *
@@ -115,4 +117,16 @@ export default class Text extends Node {
 	static fromJSON( json ) {
 		return new Text( json.data, json.attributes );
 	}
+
+	// @if CK_DEBUG_ENGINE // toString() {
+	// @if CK_DEBUG_ENGINE // 	return `#${ this.data }`;
+	// @if CK_DEBUG_ENGINE // }
+
+	// @if CK_DEBUG_ENGINE // logExtended() {
+	// @if CK_DEBUG_ENGINE // 	console.log( `ModelText: ${ this }, attrs: ${ convertMapToStringifiedObject( this.getAttributes() ) }` );
+	// @if CK_DEBUG_ENGINE // }
+
+	// @if CK_DEBUG_ENGINE // log() {
+	// @if CK_DEBUG_ENGINE // 	console.log( 'ModelText: ' + this );
+	// @if CK_DEBUG_ENGINE // }
 }
